Move errorMessages and styles out of Splash
These are not used on Splash and should be directly under
src/components.
