+++ static-site/src/components/splash/styles.jsx
@@ -197,10 +197,6 @@ const ButtonContainer = styled.button`
   }
 `;
 
-export const StrongerText = styled.span`
-  font-weight: 500;
-`;
-
 export const Button = ({to, children, target, rel}) => (
   <a href={to} target={target} rel={rel}>
     <ButtonContainer>
@@ -208,14 +204,3 @@ export const Button = ({to, children, target, rel}) => (
     </ButtonContainer>
   </a>
 );
-
-export const FixedBanner = styled.div`
-  left: 0px;
-  width: 100%;
-  height: 10%;
-  background-color: ${(props) => props.$backgroundColor};
-  font-size: 18px;
-  padding: 25px 0px 25px 0px;
-  margin: 25px 0px 25px 0px;
-  text-align: center;
-`;
