+++ static-site/src/sections/community-repo-page.jsx
@@ -5,7 +5,7 @@ import * as splashStyles from "../components/splash/styles";
 import DatasetSelect from "../components/Datasets/dataset-select";
 import GenericPage from "../layouts/generic-page";
 import { fetchAndParseJSON } from "../util/datasetsHelpers";
-import { ErrorBanner } from "../components/splash/errorMessages";
+import { ErrorBanner } from "../components/errorMessages";
 import SourceInfoHeading from "../components/sourceInfoHeading";
 
 class Index extends React.Component {
