+++ static-site/src/sections/individual-group-page.jsx
@@ -7,7 +7,7 @@ import DatasetSelect from "../components/Datasets/dataset-select";
 import GenericPage from "../layouts/generic-page";
 import { fetchAndParseJSON } from "../util/datasetsHelpers";
 import SourceInfoHeading from "../components/sourceInfoHeading";
-import { ErrorBanner } from "../components/splash/errorMessages";
+import { ErrorBanner } from "../components/errorMessages";
 import { canUserEditGroupSettings } from "./group-settings-page";
 
 class Index extends React.Component {
